controller/requirements.txt
paho-mqtt>=1.6.0
PyYAML>=6.0